feat: add style to notifications
