+++ cdk-import-service/pro.csv
@@ -1,3 +1,3 @@
 title,description,price,count
-"Product 1","Description 1",10.99,100
-"Product 2","Description 2",20.50,50
+"Product 3","Description 1",120,100
+"Product 4","Description 2",20.50,50
