+++ cdk-import-service/test.csv
@@ -1,7 +1,7 @@
 title,description,price,count
-"Product 3","Description 3",10.99,100
-"Product 4","Description 4",20.50,50
-"Product 5","Description 5",10.99,100
+"Product 5","Description 5",100.99,100
 "Product 6","Description 6",20.50,50
-"Product 7","Description 7",10.99,100
+"Product 7","Description 7",100.99,100
 "Product 8","Description 8",20.50,50
+"Product 1","Description 1",100.99,100
+"Product 2","Description 2",1740,5
