+++ cdk-product-service/bin/cdk-product-service.ts
@@ -4,4 +4,8 @@ import { CdkProductServiceStack } from '../lib/cdk-product-service-stack';
 
 const app = new cdk.App();
 new CdkProductServiceStack(app, 'CdkProductServiceStack', {
+    env: {
+        account: process.env.CDK_DEFAULT_ACCOUNT,
+        region: process.env.CDK_DEFAULT_REGION,
+      },
 });
