+++ cdk-product-service/dynamodb/createItems.ts
@@ -2,7 +2,7 @@
 import { DynamoDBDocumentClient, TransactWriteCommand , TransactWriteCommandInput} from "@aws-sdk/lib-dynamodb";
 import client from "../lambda/middleware/middleware";
 import { faker } from '@faker-js/faker';
-import {addProductTranscript} from "../lambda/request/constans";
+import {addProductTranscript} from "../lambda/request/transit/transit";
 const docClient = DynamoDBDocumentClient.from(client);
 
 function generateData(count: number) {
